split transcoder out
